requirements.txt
streamlit>=1.34
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.121.0
openai>=1.30.1
apscheduler>=3.10.4
requests>=2.31.0
